+++ prepare_common_snps.sh
@@ -88,14 +88,16 @@ if [ ! -f .${GPP_DIR}/${GPP_FILE} ]; then
     wget -O .${GPP_DIR}/${GPP_FILE}.idx ${GPP_LINK}.idx
 
     echo "Filter variants with allele frequency (AF) ≥ 0.01 to be used for variant filtering ..."
-    if [ "$1" == "local" ]; then
+    if [ "$1" == "local" ] && command -v docker &> /dev/null; then
         echo "Docker found. Running with Docker..."
         docker run --rm -w ${GPP_DIR} -v .${DATA_DIR}:${DATA_DIR} community.wave.seqera.io/library/bcftools_gatk4:d89f6490b3de65be \
             bcftools view -i 'INFO/AF>=0.01' ${GPP_DIR}/${GPP_FILE} -o ${GPP_DIR}/${GPP_FILE%.*}.af01.vcf
 
-    elif [ "$1" == "hpc" ]; then
+    elif [ "$1" == "hpc" ] && command -v singularity &> /dev/null; then
         echo "Singularity found. Running with Singularity..."
         singularity exec --pwd ${GPP_DIR} --bind .${DATA_DIR}:${DATA_DIR} docker://community.wave.seqera.io/library/bcftools_gatk4:d89f6490b3de65be \
             bcftools view -i 'INFO/AF>=0.01' ${GPP_DIR}/${GPP_FILE} -o ${GPP_DIR}/${GPP_FILE%.*}.af01.vcf
+    else
+        echo "No supported containerization tool matching argument '$1' was found in the environment."
     fi
 fi
